Changed type of retunvalue from uint16 to uint8

+++ include/Uhrtypes/DE11x11.frame.hpp
@@ -117,7 +117,7 @@ public:
 
     //------------------------------------------------------------------------------
 
-    virtual const uint16_t rowsWordMatrix() override { return 11; };
+    virtual const uint8_t rowsWordMatrix() override { return 11; };
 
     //------------------------------------------------------------------------------
 
@@ -125,7 +125,7 @@ public:
 
     //------------------------------------------------------------------------------
 
-    virtual const uint16_t numPixelsFrameMatrix() override { return 48; };
+    virtual const uint8_t numPixelsFrameMatrix() override { return 48; };
 
     //------------------------------------------------------------------------------
